:truck: Move interfaces to shared folder
