+++ src/Changelogs/AuthManager.ts
@@ -1,5 +1,5 @@
 import { UserToken } from '../types';
-import { DataStore } from '../interfaces/DataStore';
+import { DataStore } from '../shared/interfaces/DataStore';
 import { ResourceError, ResourceErrorReason } from '../errors';
