+++ src/Changelogs/ChangelogInteractor.spec.ts
@@ -1,4 +1,4 @@
-import { DataStore } from '../interfaces/DataStore';
+import { DataStore } from '../shared/interfaces/DataStore';
 import { MockDataStore } from '../tests/mock-drivers/MockDataStore';
 import { MOCK_OBJECTS } from '../tests/mocks';
 import {
