+++ src/FileManager/S3Driver.ts
@@ -7,8 +7,8 @@ import {
   CompletedPart,
   CompletedPartList,
   FileUpload,
-} from '../interfaces/FileManager';
-import { FileManager } from '../interfaces/interfaces';
+} from '../shared/interfaces/FileManager';
+import { FileManager } from '../shared/interfaces/interfaces';
 import { AWS_SDK_CONFIG } from './aws-sdk.config';
 
 AWS.config.credentials = AWS_SDK_CONFIG.credentials;
