+++ src/LearningObjects/LearningObjectRouteHandler.spec.ts
@@ -7,7 +7,7 @@ import * as bodyParser from 'body-parser';
 import * as supertest from 'supertest';
 import { MockLibraryDriver } from '../tests/mock-drivers/MockLibraryDriver';
 import { MockS3Driver } from '../tests/mock-drivers/MockS3Driver';
-import { LibraryCommunicator, FileManager } from '../interfaces/interfaces';
+import { LibraryCommunicator, FileManager } from '../shared/interfaces/interfaces';
 import * as cookieParser from 'cookie-parser';
 import { processToken, handleProcessTokenError } from '../middleware';
