+++ src/LearningObjects/LearningObjectRouteHandler.ts
@@ -1,7 +1,7 @@
 import { Request, Response, Router } from 'express';
 import { mapErrorToResponseData, ResourceErrorReason } from '../errors';
-import { DataStore } from '../interfaces/DataStore';
-import { FileManager, LibraryCommunicator } from '../interfaces/interfaces';
+import { DataStore } from '../shared/interfaces/DataStore';
+import { FileManager, LibraryCommunicator } from '../shared/interfaces/interfaces';
 import { UserToken } from '../types';
 import * as LearningObjectInteractor from './LearningObjectInteractor';
 import { LearningObject } from '../entity';
