+++ src/shared/interfaces/DataStore.ts
@@ -1,11 +1,11 @@
 import { CompletedPart, MultipartFileUploadStatus } from './FileManager';
-import { LearningObjectUpdates } from '../types';
-import { LearningOutcomeDatastore } from '../LearningOutcomes/LearningOutcomeInteractor';
-import { LearningObjectStatDatastore } from '../LearningObjectStats/LearningObjectStatsInteractor';
-import { CollectionDataStore } from '../Collections/CollectionDataStore';
-import { ChangeLogDocument } from '../types/changelog';
-import { LearningObject, User, Collection } from '../entity';
-import { Submission } from '../LearningObjectSubmission/types/Submission';
+import { LearningObjectUpdates } from '../../types';
+import { LearningOutcomeDatastore } from '../../LearningOutcomes/LearningOutcomeInteractor';
+import { LearningObjectStatDatastore } from '../../LearningObjectStats/LearningObjectStatsInteractor';
+import { CollectionDataStore } from '../../Collections/CollectionDataStore';
+import { ChangeLogDocument } from '../../types/changelog';
+import { LearningObject, User, Collection } from '../../entity';
+import { Submission } from '../../LearningObjectSubmission/types/Submission';
 
 export interface DataStore
   extends LearningOutcomeDatastore,
