+++ src/shared/interfaces/FileManager.ts
